Started Offline Mode
Started working on an offline mode to prevent the app from crashing while not connected to the internet but still allow the user to play the game.
Offline mode will be enabled automatically when there is no connection to the Gamemaker Server.
Offline mode disables login, updates, Highscores, and Achievements.